Move local mods into its own crate+trait
