+++ src/libs/ml2_mods/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod constants;
 pub mod data;
+pub mod local;
 pub mod manager;
 pub mod spelunkyfyi;
